+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 `1.0.1`
 
+- Extracted `ParameterFunction` inner class from `RequestParameters` class.
+- 
+
 ---
 
 `1.0.0`